move spec_helper customizations to other files
Because PDK wants to control spec_helper.rb, move our customizations
elsewhere. Load these customizations the same way that PDK spec_helper
does.
